fix: inaccurate variable naming
The storage variable `baseDecimals` and a function parameter
`quoteDecimal` in `UniswapTwapOracle` are inaccurate and confusing.
They're actually referring to decimal places of quote tokens instead of
base tokens. The current contract logic is correct: it's just a naming
issue.

Here we demonstrate why using quote token decimals is correct.
The function `getPairPrice` calculates the price of the base
token denominated in quote token. The immediate raw price
(`rawUniswapPriceMantissa`) is thus subject to the quote token decimals.
At the end of the function, this raw price is adjusted to the desired
output decimals of `PRICE_DECIMALS`. Thus, the decimal setting to be
compared to here should be the quote token instead.